added sql files

+++ src/main/java/com/talyas/reactiveBloodBank/repositories/DonorRepository.java
@@ -7,6 +7,6 @@ import org.springframework.stereotype.Repository;
 import reactor.core.publisher.Flux;
 
 @Repository
-public interface DonorRepository extends ReactiveCrudRepository <Donor, Integer> {
+public interface DonorRepository extends ReactiveCrudRepository <Donor, Long> {
     Flux<Donor> findAllByBloodType(EBloodType bloodType);
 }
